requirements.txt
selenium==4.13.0
openpyxl==3.1.2
pandas==2.1.1
python-dotenv==1.0.0